remove int casts

+++ examples/polybench_hls/adi.py
@@ -85,7 +85,7 @@ def adi(ttype, TSTEPS, N):
                 p[i, 0] = 0
                 q[i, 0] = v[0, i]
                 for j in range(1, N - 1):
-                    p[i, j] = int(-c / ((a * p[i, j - 1] + b)))
+                    p[i, j] = (-c / ((a * p[i, j - 1] + b)))
                     q[i, j] = ((
                         -d * u[j, i - 1]
                         + (1 + 2 * d) * u[j, i]
@@ -102,7 +102,7 @@ def adi(ttype, TSTEPS, N):
                 p[i, 0] = 0
                 q[i, 0] = u[i, 0]
                 for j in range(1, N - 1):
-                    p[i, j] = int(-f / (d * p[i, j - 1] + e))
+                    p[i, j] = (-f / (d * p[i, j - 1] + e))
                     q[i, j] = ((
                         -a * v[i - 1, j]
                         + (1.0 + 2.0 * a) * v[i, j]
